bcm2835-dma: suppress warning
Suppress the following warning by casting the pointer to and uintptr_t
before to u32:

  drivers/dma/bcm2835-dma.c: In function ‘bcm2835_dma_probe’:
  drivers/dma/bcm2835-dma.c:1179:16: warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]
    chan_start = ((u32)base / BCM2835_DMA_CHAN_SIZE) & 0xf;
                  ^

+++ drivers/dma/bcm2835-dma.c
@@ -1178,7 +1178,7 @@ static int bcm2835_dma_probe(struct platform_device *pdev)
 		return PTR_ERR(base);
 
 	/* The set of channels can be split across multiple instances. */
-	chan_start = ((u32)base / BCM2835_DMA_CHAN_SIZE) & 0xf;
+	chan_start = ((u32)(uintptr_t)base / BCM2835_DMA_CHAN_SIZE) & 0xf;
 	base -= BCM2835_DMA_CHAN(chan_start);
 	chan_count = resource_size(res) / BCM2835_DMA_CHAN_SIZE;
 	chan_end = min(chan_start + chan_count,
